refactor(backend): use schematic for env vars
