+++ apps/backend/src/importer/mod.rs
@@ -14,7 +14,6 @@ use crate::{
     audio_books::resolver::AudioBooksService,
     background::ImportMedia,
     books::resolver::BooksService,
-    config::ImporterConfig,
     entities::{media_import_report, prelude::MediaImportReport},
     media::resolver::{MediaDetails, MediaService, ProgressUpdate, ProgressUpdateAction},
     migrator::{MediaImportSource, MetadataLot},
@@ -254,12 +253,7 @@ impl ImporterService {
         self.misc_service.media_import_reports(user_id).await
     }
 
-    pub async fn import_from_source(
-        &self,
-        user_id: i32,
-        input: DeployImportInput,
-        config: &ImporterConfig,
-    ) -> Result<()> {
+    pub async fn import_from_source(&self, user_id: i32, input: DeployImportInput) -> Result<()> {
         let db_import_job = self
             .misc_service
             .start_import_job(user_id, input.source)
@@ -268,9 +262,7 @@ impl ImporterService {
             MediaImportSource::MediaTracker => {
                 media_tracker::import(input.media_tracker.unwrap()).await?
             }
-            MediaImportSource::Goodreads => {
-                goodreads::import(input.goodreads.unwrap(), config).await?
-            }
+            MediaImportSource::Goodreads => goodreads::import(input.goodreads.unwrap()).await?,
         };
         for (idx, item) in import.media.iter().enumerate() {
             tracing::trace!(
